Remove cruft (#12)
* Remove glimr API unit tests

I put these in place for the original port from the prototype. After
some deliberation, I decided to remove them in favour of feature and
integration specs, which offer more sweeping coverage and make the code
easier to re-think and refactor.

* Remove unused code from the glimr api

The plan is to evenutally refactor this code to reduce its complexity
somewhat.  For now, however, I am removing a few things that are not
currently used anywhwere.

* Move the unit test for network failure

This should make refactoring simpler.

* Continually check that glimr is available

Previously, we were only checking this at the start of the run. This did
not account for the possibility that glimr could go away at any point in
the 'transaction'. The user would not have been notified until they
attempted to complete a payment.  This will stop them as move between
'states'/pages.
