+++ app/services/glimr.rb
@@ -2,7 +2,6 @@ require 'glimr/api'
 require 'glimr/requests/base'
 require 'glimr/requests/case_fees'
 require 'glimr/requests/fee_paid'
-require 'glimr/responses/api_error'
 require 'glimr/responses/case_fees'
 require 'glimr/responses/case_not_found'
 require 'glimr/responses/fee_payment'
