--- app/services/glimr/responses/api_error.rb
@@ -1,21 +1,0 @@
-module Glimr
-  module Responses
-    class ApiError
-      def initialize(error)
-        @error = error
-      end
-
-      def error?
-        true
-      end
-
-      def error_code
-        @error.class
-      end
-
-      def error_message
-        @error.message
-      end
-    end
-  end
-end
