+++ app/services/glimr/responses/fee_payment.rb
@@ -8,10 +8,6 @@ module Glimr
       def error?
         false
       end
-
-      def fee_transaction_id
-        @glimr_response['feeTransactionId']
-      end
     end
   end
 end
