+++ config/initializers/high_voltage.rb
@@ -1,3 +1,3 @@
 HighVoltage.configure do |config|
-  config.routes = false
+  config.home_page = 'start'
 end
