+++ config/routes.rb
@@ -1,8 +1,4 @@
 Rails.application.routes.draw do
-  get '/pages/*id' => 'pages#show', as: :page, format: false
-
-  root to: 'pages#show', id: 'start'
-
   resources :case_requests,
     only: [:new, :create],
     path_names: { new: '' }
