+++ spec/rails_helper.rb
@@ -22,6 +22,9 @@ RSpec.configure do |config|
   config.before(:each) do
     I18n.locale = I18n.default_locale
     DatabaseCleaner.strategy = :transaction
+    stub_request(:post, 'https://glimr-test.dsd.io/glimravailable').
+      with(:headers => {'Accept' => 'application/json'}).
+      to_return(status: 200, body: '{"glimrAvailable":"yes"}')
   end
 
   config.before(:each, js: true) do
